property file is not required

+++ chaco/generate_job.py
@@ -17,7 +17,7 @@ class Generate_Job(Command):
     def get_parser(self, prog_name):
         parser = super(Generate_Job, self).get_parser(prog_name)
         parser.add_argument('--jobyaml', required=True)
-        parser.add_argument('--propertyfile', required=True)
+        parser.add_argument('--propertyfile')
         parser.add_argument('--outputdir', required=True)
         return parser
 
@@ -25,10 +25,6 @@ class Generate_Job(Command):
         jobyaml = parsed_args.jobyaml
         if os.path.exists(jobyaml) == False:
             raise Exception("%s doesn't exist" % (jobyaml))
-
-        propertyfile = parsed_args.propertyfile
-        if os.path.exists(propertyfile) == False:
-            raise Exception("%s doesn't exist" % (propertyfile))
 
         outputdir = parsed_args.outputdir
         if os.path.exists(outputdir):
@@ -39,10 +35,16 @@ class Generate_Job(Command):
             y = yaml.load(f)
             zipfile_path = os.path.join(outputdir, os.path.basename(f.name) + ".zip")
             zip = zipfile.ZipFile(zipfile_path, "w", zipfile.ZIP_DEFLATED)
-            shutil.copy(propertyfile, outputdir)
-            p = os.path.join(outputdir, os.path.basename(propertyfile))
-            zip.write(p)
-            os.remove(p)
+
+            if parsed_args.propertyfile:
+                propertyfile = parsed_args.propertyfile
+                if os.path.exists(propertyfile) == False:
+                    raise Exception("%s doesn't exist" % (propertyfile))
+                shutil.copy(propertyfile, outputdir)
+                p = os.path.join(outputdir, os.path.basename(propertyfile))
+                zip.write(p)
+                os.remove(p)
+
             for flow_name, flow_content in y.items():
                 self.log.debug("flow_name=%s, flow_content=%s" % (flow_name, flow_content))
                 flow_dir = os.path.join(outputdir, flow_name)
